fix(epic-web): keep mysql2 out of client bundle for contributor page
The previous commit added a top-level `mysql2/promise` import to
lib/contributors.ts, but that module is imported by client components
(contributor-bio, share), so the Node-only `net`/`tls` deps were pulled into
the client bundle and broke the production build ("Can't resolve 'net'").

Follow the existing `.server.ts` convention (cf. lib/tips.server.ts): move
getContributorResources and the Course Builder MySQL query into a new
lib/contributors.server.ts, leaving lib/contributors.ts free of mysql2 so it
stays safe to import from the client. The page imports getContributorResources
from the server module.

Verified: production build compiles with no net/tls resolution errors.
