+++ apps/epic-web/src/lib/contributors.ts
@@ -1,11 +1,6 @@
 import {sanityClient} from '../utils/sanity-client'
 import groq from 'groq'
 import z from 'zod'
-import * as mysql from 'mysql2/promise'
-
-const access: mysql.ConnectionOptions = {
-  uri: process.env.COURSE_BUILDER_DATABASE_URL,
-}
 
 export const ContributorRoleSchema = z.enum([
   'author',
@@ -129,143 +124,6 @@ export const ContributorResourcesSchema = z.array(ContributorResourceSchema)
 
 export type ContributorResource = z.infer<typeof ContributorResourceSchema>
 
-// Articles published from Course Builder live in the Course Builder MySQL DB,
-// not in Sanity. Mirror lib/articles.ts (getAllArticles) and pull them in, keyed
-// by the contributor's userId (which equals `createdById` on a Course Builder
-// post). Without this, Course Builder articles render as standalone article
-// pages but never appear on the author's contributor page.
-const getContributorCourseBuilderArticles = async (
-  userId: string,
-): Promise<ContributorResource[]> => {
-  const connection = await mysql.createConnection(access)
-
-  try {
-    const [rows] = await connection.execute(
-      `SELECT * FROM zEW_ContentResource
-       WHERE JSON_EXTRACT(fields, "$.postType") = 'article'
-         AND JSON_EXTRACT(fields, "$.state") = 'published'
-         AND createdById = ?`,
-      [userId],
-    )
-
-    if (!Array.isArray(rows)) return []
-
-    return Promise.all(
-      (rows as any[]).map(async (row): Promise<ContributorResource> => {
-        const fields =
-          typeof row.fields === 'string'
-            ? JSON.parse(row.fields)
-            : row.fields || {}
-
-        // Find the article's video (if any) for the contributor-page thumbnail.
-        let muxPlaybackId: string | null = null
-        try {
-          const [videoRows] = await connection.execute(
-            `SELECT resource.fields
-             FROM zEW_ContentResourceResource crr
-             JOIN zEW_ContentResource resource ON crr.resourceId = resource.id
-             WHERE crr.resourceOfId = ?
-               AND resource.type = 'videoResource'
-               AND crr.deletedAt IS NULL
-               AND resource.deletedAt IS NULL
-             LIMIT 1`,
-            [row.id],
-          )
-
-          if (Array.isArray(videoRows) && videoRows.length > 0) {
-            const videoFields =
-              typeof (videoRows[0] as any).fields === 'string'
-                ? JSON.parse((videoRows[0] as any).fields)
-                : (videoRows[0] as any).fields || {}
-            muxPlaybackId = videoFields.muxPlaybackId || null
-          }
-        } catch (error) {
-          console.error(
-            '[getContributorResources] Error fetching video for article:',
-            error,
-          )
-        }
-
-        return {
-          _id: row.id,
-          _type: 'article',
-          _updatedAt: row.updatedAt
-            ? new Date(row.updatedAt).toISOString()
-            : '',
-          _createdAt: row.createdAt
-            ? new Date(row.createdAt).toISOString()
-            : '',
-          title: fields.title || '',
-          slug: fields.slug || '',
-          description: fields.description || null,
-          moduleType: null,
-          image: null,
-          muxPlaybackId,
-        }
-      }),
-    )
-  } finally {
-    await connection.end()
-  }
-}
-
-export const getContributorResources = async (
-  id: string,
-  userId?: string | null,
-): Promise<ContributorResource[] | null> => {
-  const sanityResources = await sanityClient.fetch(
-    groq`*[$id in contributors[].contributor._ref && _type in ["article", "tip", "module", "talk"] && state == 'published'] | order(_createdAt desc) {
-            _id,
-            _type,
-            _updatedAt,
-            _createdAt,
-            title,
-            description,
-            summary,
-            "slug": slug.current,
-            "moduleType": moduleType,
-            "image": coalesce(image.asset->url, image.secure_url),
-            "muxPlaybackId": resources[@->._type == 'videoResource'][0]-> muxAsset.muxPlaybackId,
-    }`,
-    {id},
-  )
-
-  let courseBuilderResources: ContributorResource[] = []
-  if (userId) {
-    try {
-      courseBuilderResources = await getContributorCourseBuilderArticles(userId)
-    } catch (error) {
-      console.error(
-        '[getContributorResources] Error fetching Course Builder articles:',
-        error,
-      )
-    }
-  }
-
-  // Merge both sources, drop duplicate slugs (an article may exist in both
-  // after a migration), and keep the newest-to-oldest ordering of the page.
-  const seen = new Set<string>()
-  const merged = [
-    ...(Array.isArray(sanityResources) ? sanityResources : []),
-    ...courseBuilderResources,
-  ]
-    .filter((resource: any) => {
-      const slug = resource?.slug
-      if (!slug) return true
-      if (seen.has(slug)) return false
-      seen.add(slug)
-      return true
-    })
-    .sort(
-      (a: any, b: any) =>
-        new Date(b._createdAt).getTime() - new Date(a._createdAt).getTime(),
-    )
-
-  const result = ContributorResourcesSchema.safeParse(merged)
-
-  if (result.success) {
-    return result.data
-  } else {
-    return null
-  }
-}
+// `getContributorResources` lives in lib/contributors.server.ts because it
+// queries the Course Builder MySQL DB (mysql2). Keeping that import out of this
+// module is what keeps it safe to import from client components.
